feat(builtin): split node_loader into node_patcher
