+++ internal/node/test/child_process/lib/BUILD.bazel
@@ -1,0 +1,9 @@
+load("//internal/js_library:js_library.bzl", "js_library")
+
+package(default_visibility = ["//internal:__subpackages__"])
+
+js_library(
+    name = "lib",
+    srcs = ["index.js"],
+    module_name = "lib",
+)
