+++ internal/node/test/child_process/lib/index.js
@@ -1,0 +1,1 @@
+console.log('LIB')
