+++ internal/node/test/child_process/worker-main.js
@@ -1,0 +1,4 @@
+const childProcess = require('child_process');
+console.log('WORKER-MAIN');
+require('lib');
+new childProcess.fork(__dirname + '/worker-thread.js');
