+++ internal/node/test/worker_threads/lib/index.js
@@ -1,0 +1,1 @@
+console.log('LIB')
