+++ internal/node/test/worker_threads/worker-main.js
@@ -1,0 +1,4 @@
+const workerThreads = require('worker_threads');
+console.log('WORKER-MAIN');
+require('lib');
+new workerThreads.Worker(__dirname + '/worker-thread.js');
