+++ internal/node/test/worker_threads/worker-thread.js
@@ -1,0 +1,2 @@
+console.log('WORKER THREAD');
+require('lib');
